% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_recommendations.R
\name{create_rec_playlist}
\alias{create_rec_playlist}
\title{Create a playlist}
\usage{
create_rec_playlist(
  playlist_name,
  authorization = get_spotify_authorization_code()
)
}
\arguments{
\item{playlist_name}{The name of the new playlist.}

\item{authorization}{The authorization token or code for the user's account.}
}
\value{
A list containing the url and playlist id for the new playlist.
}
\description{
Creates a new, empty playlist for the user on Spotify.
}
