% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_recommendations.R
\name{extract_artist_id}
\alias{extract_artist_id}
\title{Extract Artist ID}
\usage{
extract_artist_id(artists)
}
\arguments{
\item{artists}{A vector containing artist information.}
}
\value{
A vector of artist IDs.
}
\description{
Extract Artist ID
}
