% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_playlists_songs.R
\name{filter_playlists}
\alias{filter_playlists}
\title{Filters a dataframe of playlists}
\usage{
filter_playlists(all_playlists, playlist_names = NULL)
}
\arguments{
\item{all_playlists}{a dataframe containing names and IDs of playlists}

\item{playlist_names}{a character vector containing the playlist names, defaults to all}
}
\value{
dataframe containing the names and IDs of playlists
}
\description{
Filters a dataframe of playlists
}
