% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_top_genres.R
\name{get_all_artists}
\alias{get_all_artists}
\title{Get Artist information}
\usage{
get_all_artists(saved_tracks)
}
\arguments{
\item{saved_tracks}{A dataframe containing metadata of
the user's saved tracks.}
}
\value{
A list containing metadata of all the artists.
}
\description{
A function that iteratively calls the Spotify API #' to get
information on all the artists present in a list of tracks.
}
