% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_playlists_songs.R
\name{get_all_playlists}
\alias{get_all_playlists}
\title{Get all the names and ids of a user's playlists}
\usage{
get_all_playlists(authorization)
}
\arguments{
\item{authorization}{an authorization token for the Spotify API}
}
\value{
dataframe containing the names and IDs of playlists
}
\description{
Get all the names and ids of a user's playlists
}
