% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_releases_by_continent.R
\name{get_map}
\alias{get_map}
\title{Get the country list by continent}
\usage{
get_map()
}
\value{
A list containing all countries, by continent name
}
\description{
Get the country list by continent
}
