% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_releases_by_continent.R,
%   R/spotify_assistant.R
\name{get_new_releases_by_continent}
\alias{get_new_releases_by_continent}
\title{Get names of new releases by continent}
\usage{
get_new_releases_by_continent(country_code, n_limit)

get_new_releases_by_continent(country_code, n_limit)
}
\arguments{
\item{n_limit}{Number of new releases to return}

\item{continent}{: continent name (i.e. Asia, Europe, Oceania, Americas, Africa)}

\item{limit}{: max number of albums to extract}

\item{country_cde}{A two digit string that}
}
\value{
A list of titles of new releases in String from the corresponding continent

List of new track releases in the given continent
}
\description{
This is a helper function that allows users to get new releases
by continent. It goes through all the countries in the region
and summaries the info of all releases. If we specify a n_limit
smaller than the total number of all releases available, it returns
the first n_limit releases by default
}
\examples{
\dontrun{
get_new_releases_by_continent('Asia', 5)
get_new_releases_by_continent('Americas', 1)
}
get_new_releases_by_continent('Asia', 10)
get_new_releases_by_continent('Europe', 15)
}
