% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_recommendations.R
\name{get_new_songs}
\alias{get_new_songs}
\title{Get New Songs}
\usage{
get_new_songs(
  seed_type,
  seeds,
  num_songs = 10,
  authorization = get_spotify_access_token()
)
}
\arguments{
\item{seed_type}{Either 'artists' or 'genres'. Default is 'artists'.}

\item{seeds}{A vector of artist or track ID's.}

\item{num_songs}{The number of recommended songs to return.
Must be between 1 and 100 (inclusive).}

\item{authorization}{The authorization token or code for the user's account.}
}
\value{
A vector of track uri's for identifying specific tracks.
}
\description{
Returns a specified number of recommended songs from Spotify.
}
