% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_playlists_songs.R
\name{get_one_playlist_songs}
\alias{get_one_playlist_songs}
\title{Get all the songs from a playlist}
\usage{
get_one_playlist_songs(playlist_id)
}
\arguments{
\item{playlist_id}{a character string containing the ID for a Spotify playlist}
}
\value{
character vector containing song names
}
\description{
Get all the songs from a playlist
}
