% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_playlists_songs.R,
%   R/spotify_assistant.R
\name{get_playlists_songs}
\alias{get_playlists_songs}
\title{Get all the songs from a user's playlists}
\usage{
get_playlists_songs(playlists = NULL)

get_playlists_songs(playlists = NULL)
}
\arguments{
\item{playlists}{a character vector containing the playlist names, defaults to all}

\item{playlist_names}{a character vector containing the playlist names, defaults to all}
}
\value{
list containing a vector of song names from the playlists

list containing a vector of song names from the playlists
}
\description{
A wrapper function that gets the names of all songs contained in one or more
of a user's owned and saved public or collaborative playlists.
By default, all songs from valid playlists are returned,
but playlists can also be specified by name.

A wrapper function that gets the names of all songs contained in one or more
of a user's owned and saved playlists. By default, all songs from all playlists
are returned, but playlists can also be specified by name.
}
\examples{
get_playlists_songs()
get_playlists_songs('night drives')
get_playlists_songs(c('night drives', 'boss rush'))
get_playlists_songs()
get_playlists_songs('night drives')
get_playlists_songs(c('night drives', 'boss rush'))
}
