% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_releases_by_continent.R
\name{get_releases}
\alias{get_releases}
\title{Get all releases by continent}
\usage{
get_releases(continent, limit)
}
\arguments{
\item{continent}{continent name as a String}

\item{limit}{number of releases to extract}
}
\value{
a list containing all info of selected new releases
}
\description{
Get all releases by continent
}
