% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_top_genres.R
\name{get_saved_tracks}
\alias{get_saved_tracks}
\title{Get the user's saved tracks}
\usage{
get_saved_tracks(authorization)
}
\arguments{
\item{authorization}{an authorization token for the Spotify API}
}
\value{
A dataframe containing metadata of the user's saved tracks.
}
\description{
A function that iteratively calls the Spotify API to get a user's
saved tracks and the metadata pertaining to each track.
}
