% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_recommendations.R
\name{get_song_recommendations}
\alias{get_song_recommendations}
\title{Creates a playlist of recommended songs.}
\usage{
get_song_recommendations(playlist_name = NULL, num_songs = 10)
}
\arguments{
\item{playlist_name}{The name of the new playlist, a string. Defaults to
"Recommended Songs" with the current date.}

\item{num_songs}{The number of new songs to generate. Must be an integer
between 1 and 100 (inclusive).}
}
\description{
A wrapper function that first retrieves a user's top 3 artists,
then uses this information to generate a list of 10 recommended songs.
If the user does not have any top artist information, then genre seeds
are used instead.
A new playlist is created for the user, which contains the
recommended songs.
Prints a url link to the new playlist on Spotify.
}
\examples{
get_song_recommendations()
}
