% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_song_recommendations.R
\name{get_top_artists}
\alias{get_top_artists}
\title{Get Top Artists}
\usage{
get_top_artists()
}
\value{
A data frame containing artist information.
}
\description{
Returns the current user's top artists from Spotify.
}
