% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_top_genres.R
\name{get_top_genres}
\alias{get_top_genres}
\title{Get the top genres}
\usage{
get_top_genres(artist_information)
}
\arguments{
\item{artist_information}{A list containing metadata of all the artists.}
}
\value{
A vector containing the top 5 occurrences.
}
\description{
A function that that iterates through a list of genres, counts the number of
occurrences of each genres and returns the top 5 occurrences.
}
