% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_users_top_genres.R
\name{get_users_top_genres}
\alias{get_users_top_genres}
\title{Finds the top 5 genres from a user's saved tracks}
\usage{
get_users_top_genres()
}
\value{
Character vector containing the top 5 genres of music in the User's saved library.
}
\description{
A wrapper function that goes through a user's entire saved library.
For each song, it checks all the artist involved and the genre of music
they compose. It returns the top 5 most commonly occurring genres.
}
