% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_releases_by_continent.R
\name{region_not_available}
\alias{region_not_available}
\title{Check if the country code is valid}
\usage{
region_not_available(code)
}
\arguments{
\item{code}{country code to look up}
}
\value{
If spotify is available in the country or not
}
\description{
Check if the country code is valid
}
