% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_releases_by_continent.R
\name{get_new_releases_by_continent}
\alias{get_new_releases_by_continent}
\title{Get names of new releases by continent}
\usage{
get_new_releases_by_continent(continent, limit)
}
\arguments{
\item{continent}{: continent name (i.e. Asia, Europe, Oceania, Americas, Africa)}

\item{limit}{: max number of albums to extract}
}
\value{
A list of titles of new releases in String from the corresponding continent
}
\description{
Get names of new releases by continent
}
\examples{
\dontrun{
get_new_releases_by_continent('Asia', 5)
get_new_releases_by_continent('Americas', 1)
}
}
